move install and uninstall to kubernetes app
