+++ kpops/components/streams_bootstrap/__init__.py
@@ -1,2 +1,6 @@
-from kpops.components.streams_bootstrap.producer import ProducerApp  # noqa F401
-from kpops.components.streams_bootstrap.streams_app import StreamsApp  # noqa F401
+from kpops.components.streams_bootstrap.producer.producer_app import (  # noqa F401
+    ProducerApp,
+)
+from kpops.components.streams_bootstrap.streams.streams_app import (  # noqa F401
+    StreamsApp,
+)
