+++ tests/cli/test_handlers.py
@@ -7,7 +7,6 @@ from kpops.cli.pipeline_config import PipelineConfig
 from kpops.component_handlers import ComponentHandlers
 from kpops.component_handlers.kafka_connect.connector_handler import ConnectorHandler
 from kpops.component_handlers.schema_handler.schema_handler import SchemaHandler
-from kpops.component_handlers.streams_bootstrap.app_handler import AppHandler
 from kpops.component_handlers.topic.handler import TopicHandler
 from tests.cli.resources.module import CustomSchemaProvider
 
@@ -21,10 +20,6 @@ def test_set_up_handlers_with_no_schema_handler(mocker: MockerFixture):
         kafka_rest_host="https://testhost:8082",
         schema_registry_url=None,
     )
-    app_handler_mock = mocker.patch("kpops.cli.main.AppHandler")
-    app_handler = AppHandler.from_pipeline_config(config)
-    app_handler_mock.from_pipeline_config.return_value = app_handler
-
     connector_handler_mock = mocker.patch("kpops.cli.main.ConnectorHandler")
     connector_handler = ConnectorHandler.from_pipeline_config(pipeline_config=config)
     connector_handler_mock.from_pipeline_config.return_value = connector_handler
@@ -36,23 +31,19 @@ def test_set_up_handlers_with_no_schema_handler(mocker: MockerFixture):
 
     expected = ComponentHandlers(
         schema_handler=None,
-        app_handler=app_handler,
         connector_handler=connector_handler,
         topic_handler=topic_handler,
     )
 
     actual_handlers = setup_handlers(MODULE, config)
 
-    app_handler_mock.from_pipeline_config.assert_called_once_with(config)
     connector_handler_mock.from_pipeline_config.assert_called_once_with(config)
 
     assert actual_handlers.schema_handler == expected.schema_handler
-    assert actual_handlers.app_handler == expected.app_handler
     assert actual_handlers.connector_handler == expected.connector_handler
     assert actual_handlers.topic_handler == expected.topic_handler
 
     assert actual_handlers.schema_handler is None
-    assert type(actual_handlers.app_handler) is AppHandler
     assert type(actual_handlers.connector_handler) is ConnectorHandler
     assert type(actual_handlers.topic_handler) is TopicHandler
 
@@ -68,10 +59,6 @@ def test_set_up_handlers_with_schema_handler(mocker: MockerFixture):
     schema_handler = SchemaHandler.load_schema_handler(MODULE, config)
     schema_handler_mock.load_schema_handler.return_value = schema_handler
 
-    app_handler_mock = mocker.patch("kpops.cli.main.AppHandler")
-    app_handler = AppHandler.from_pipeline_config(config)
-    app_handler_mock.from_pipeline_config.return_value = app_handler
-
     connector_handler_mock = mocker.patch("kpops.cli.main.ConnectorHandler")
     connector_handler = ConnectorHandler.from_pipeline_config(pipeline_config=config)
     connector_handler_mock.from_pipeline_config.return_value = connector_handler
@@ -83,7 +70,6 @@ def test_set_up_handlers_with_schema_handler(mocker: MockerFixture):
 
     expected = ComponentHandlers(
         schema_handler=schema_handler,
-        app_handler=app_handler,
         connector_handler=connector_handler,
         topic_handler=topic_handler,
     )
@@ -92,15 +78,12 @@ def test_set_up_handlers_with_schema_handler(mocker: MockerFixture):
 
     schema_handler_mock.load_schema_handler.assert_called_once_with(MODULE, config)
 
-    app_handler_mock.from_pipeline_config.assert_called_once_with(config)
     connector_handler_mock.from_pipeline_config.assert_called_once_with(config)
 
     assert actual_handlers.schema_handler == expected.schema_handler
-    assert actual_handlers.app_handler == expected.app_handler
     assert actual_handlers.connector_handler == expected.connector_handler
     assert actual_handlers.topic_handler == expected.topic_handler
 
     assert type(actual_handlers.schema_handler) is SchemaHandler
-    assert type(actual_handlers.app_handler) is AppHandler
     assert type(actual_handlers.connector_handler) is ConnectorHandler
     assert type(actual_handlers.topic_handler) is TopicHandler
