+++ tests/components/test_base_defaults_component.py
@@ -47,7 +47,6 @@ def config() -> PipelineConfig:
 def handlers() -> ComponentHandlers:
     return ComponentHandlers(
         schema_handler=MagicMock(),
-        app_handler=MagicMock(),
         connector_handler=MagicMock(),
         topic_handler=MagicMock(),
     )
