+++ tests/components/test_pipeline_component.py
@@ -19,7 +19,6 @@ class TestPipelineComponent:
         pipeline_component = PipelineComponent(
             handlers=ComponentHandlers(
                 schema_handler=MagicMock(),
-                app_handler=MagicMock(),
                 connector_handler=MagicMock(),
                 topic_handler=MagicMock(),
             ),
